moved test in a better place
